+++ src/modules/vtex/util/child-process.ts
@@ -1,16 +1,16 @@
 import { consts } from "./../../../shared/constants";
 import {
-  link_succes_log,
+  link_success_log,
   link_error_log,
-  publish_succes_log,
+  publish_success_log,
   publish_error_log
 } from "../../../shared/log-vtex";
 import { OptRunCommand, LogContent } from "../../../shared/models/global";
 const { spawn } = require("child_process");
 import log from "../../../shared/logger";
 
-export const runCommand = function(options: OptRunCommand) {
-  return new Promise(function(fulfill, reject) {
+export const runCommand = function (options: OptRunCommand) {
+  return new Promise(function (fulfill, reject) {
     log.info(`Using component in the folder: ${options.path}`);
 
     const task = spawn(`cd ${options.path} && ${options.command}`, [], {
@@ -21,7 +21,7 @@ export const runCommand = function(options: OptRunCommand) {
     task.stdout.on("data", (data: string) => {
       log.info(data.toString());
 
-      if (validateLog(options.command, data.toString(), "succes")) {
+      if (validateLog(options.command, data.toString(), "success")) {
         task.kill("SIGINT");
         fulfill(true);
       }
@@ -35,7 +35,7 @@ export const runCommand = function(options: OptRunCommand) {
     });
 
     // Método para imprimir el log de error
-    task.stderr.on("data", function(data: string) {
+    task.stderr.on("data", function (data: string) {
       log.info(data.toString());
       // Validación en caso de error
       if (validateLog(options.command, data.toString(), "error")) {
@@ -45,7 +45,7 @@ export const runCommand = function(options: OptRunCommand) {
         process.exit(1);
       }
 
-      if (validateLog(options.command, data.toString(), "succes")) {
+      if (validateLog(options.command, data.toString(), "success")) {
         task.kill("SIGINT");
         fulfill(true);
       }
@@ -60,30 +60,30 @@ export const runCommand = function(options: OptRunCommand) {
  */
 export const validateLog = (
   command: string,
-  logMesage: string,
+  logMessage: string,
   type: string
 ) => {
   if (command == consts.link) {
-    if (type == "succes") {
-      const isFind = !!link_succes_log.find((element: LogContent) =>
-        logMesage.includes(element.log)
+    if (type == "success") {
+      const isFind = !!link_success_log.find((element: LogContent) =>
+        logMessage.includes(element.log)
       );
       return isFind;
     } else if (type == "error") {
       const isFind = !!link_error_log.find((element: LogContent) =>
-        logMesage.includes(element.log)
+        logMessage.includes(element.log)
       );
       return isFind;
     }
   } else if (command == consts.publish) {
-    if (type == "succes") {
-      const isFind = !!publish_succes_log.find((element: LogContent) =>
-        logMesage.includes(element.log)
+    if (type == "success") {
+      const isFind = !!publish_success_log.find((element: LogContent) =>
+        logMessage.includes(element.log)
       );
       return isFind;
     } else if (type == "error") {
       const isFind = !!publish_error_log.find((element: LogContent) =>
-        logMesage.includes(element.log)
+        logMessage.includes(element.log)
       );
       return isFind;
     }