+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### v2.7.6 2019-07-24
+
+## Added
+
+* Change validation to generate versions, add `--last` validation
+
+* Added logic for scape the command passed on `link` or `publish` in vtex, this option is for AWS CodeBuild process.
+
+* Added `exito vtex updateDepProject` command
+
 ### v2.7.0 2019-07-22
 
 ## Changed